third_party/WebKit/LayoutTests/http/tests/eventsource/workers/eventsource-cors-non-http-expected.txt

CONSOLE ERROR: EventSource cannot load ftp://127.0.0.1/. Cross origin requests are only supported for protocol schemes: http, data, https.
CONSOLE ERROR: EventSource cannot load motd. Cross origin requests are only supported for protocol schemes: http, data, https.
CONSOLE ERROR: EventSource cannot load localhost:8080/. Cross origin requests are only supported for protocol schemes: http, data, https.
CONSOLE ERROR: EventSource cannot load tel:1234. Cross origin requests are only supported for protocol schemes: http, data, https.
[Worker] Test EventSource with non-HTTP protocol schemes in the URL.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: ../script-tests/eventsource-cors-non-http.js
PASS [Worker] es.readyState is EventSource.CLOSED
PASS [Worker] es.readyState is EventSource.CLOSED
PASS [Worker] es.readyState is EventSource.CLOSED
PASS [Worker] es.readyState is EventSource.CLOSED
PASS successfullyParsed is true

TEST COMPLETE
